mydylms-api/requirements.txt
beautifulsoup4==4.14.2
bs4==0.0.2
fastapi==0.118.3
httpx==0.28.1
pydantic==2.12.0
python-dotenv==1.1.1
python-json-logger==4.0.0
requests==2.32.5
uvicorn==0.37.0
pytest==8.4.2